third_party/blink/web_tests/external/wpt/css/css-pseudo/highlight-pseudos-computed-expected.txt

This is a testharness.js-based test.
[FAIL] getComputedStyle() for ::selection: should return an empty CSSStyleDeclaration
  assert_equals: expected "" but got "rgba(0, 0, 0, 0)"
[FAIL] getComputedStyle() for ::selection) should return an empty CSSStyleDeclaration
  assert_equals: expected "" but got "rgba(0, 0, 0, 0)"
[FAIL] getComputedStyle() for ::selection( should return an empty CSSStyleDeclaration
  assert_equals: expected "" but got "rgba(0, 0, 0, 0)"
[FAIL] getComputedStyle() for ::selection(foo) should return an empty CSSStyleDeclaration
  assert_equals: expected "" but got "rgba(0, 0, 0, 0)"
[FAIL] getComputedStyle() for ::selection() should return an empty CSSStyleDeclaration
  assert_equals: expected "" but got "rgba(0, 0, 0, 0)"
[FAIL] getComputedStyle() for :::selection should return an empty CSSStyleDeclaration
  assert_equals: expected "" but got "rgba(0, 0, 0, 0)"
[FAIL] getComputedStyle() for ::selection. should return an empty CSSStyleDeclaration
  assert_equals: expected "" but got "rgba(0, 0, 0, 0)"
[FAIL] getComputedStyle() for ::target-text: should return an empty CSSStyleDeclaration
  assert_equals: expected "" but got "rgba(0, 0, 0, 0)"
[FAIL] getComputedStyle() for ::target-text) should return an empty CSSStyleDeclaration
  assert_equals: expected "" but got "rgba(0, 0, 0, 0)"
[FAIL] getComputedStyle() for ::target-text( should return an empty CSSStyleDeclaration
  assert_equals: expected "" but got "rgba(0, 0, 0, 0)"
[FAIL] getComputedStyle() for ::target-text(foo) should return an empty CSSStyleDeclaration
  assert_equals: expected "" but got "rgba(0, 0, 0, 0)"
[FAIL] getComputedStyle() for ::target-text() should return an empty CSSStyleDeclaration
  assert_equals: expected "" but got "rgba(0, 0, 0, 0)"
[FAIL] getComputedStyle() for :::target-text should return an empty CSSStyleDeclaration
  assert_equals: expected "" but got "rgba(0, 0, 0, 0)"
[FAIL] getComputedStyle() for ::target-text. should return an empty CSSStyleDeclaration
  assert_equals: expected "" but got "rgba(0, 0, 0, 0)"
[FAIL] getComputedStyle() for ::spelling-error: should return an empty CSSStyleDeclaration
  assert_equals: expected "" but got "rgba(0, 0, 0, 0)"
[FAIL] getComputedStyle() for ::spelling-error) should return an empty CSSStyleDeclaration
  assert_equals: expected "" but got "rgba(0, 0, 0, 0)"
[FAIL] getComputedStyle() for ::spelling-error( should return an empty CSSStyleDeclaration
  assert_equals: expected "" but got "rgba(0, 0, 0, 0)"
[FAIL] getComputedStyle() for ::spelling-error(foo) should return an empty CSSStyleDeclaration
  assert_equals: expected "" but got "rgba(0, 0, 0, 0)"
[FAIL] getComputedStyle() for ::spelling-error() should return an empty CSSStyleDeclaration
  assert_equals: expected "" but got "rgba(0, 0, 0, 0)"
[FAIL] getComputedStyle() for :::spelling-error should return an empty CSSStyleDeclaration
  assert_equals: expected "" but got "rgba(0, 0, 0, 0)"
[FAIL] getComputedStyle() for ::spelling-error. should return an empty CSSStyleDeclaration
  assert_equals: expected "" but got "rgba(0, 0, 0, 0)"
[FAIL] getComputedStyle() for ::grammar-error: should return an empty CSSStyleDeclaration
  assert_equals: expected "" but got "rgba(0, 0, 0, 0)"
[FAIL] getComputedStyle() for ::grammar-error) should return an empty CSSStyleDeclaration
  assert_equals: expected "" but got "rgba(0, 0, 0, 0)"
[FAIL] getComputedStyle() for ::grammar-error( should return an empty CSSStyleDeclaration
  assert_equals: expected "" but got "rgba(0, 0, 0, 0)"
[FAIL] getComputedStyle() for ::grammar-error(foo) should return an empty CSSStyleDeclaration
  assert_equals: expected "" but got "rgba(0, 0, 0, 0)"
[FAIL] getComputedStyle() for ::grammar-error() should return an empty CSSStyleDeclaration
  assert_equals: expected "" but got "rgba(0, 0, 0, 0)"
[FAIL] getComputedStyle() for :::grammar-error should return an empty CSSStyleDeclaration
  assert_equals: expected "" but got "rgba(0, 0, 0, 0)"
[FAIL] getComputedStyle() for ::grammar-error. should return an empty CSSStyleDeclaration
  assert_equals: expected "" but got "rgba(0, 0, 0, 0)"
[FAIL] getComputedStyle() for ::highlight(foo): should return an empty CSSStyleDeclaration
  assert_equals: expected "" but got "rgba(0, 0, 0, 0)"
[FAIL] getComputedStyle() for ::highlight(foo)) should return an empty CSSStyleDeclaration
  assert_equals: expected "" but got "rgba(0, 0, 0, 0)"
[FAIL] getComputedStyle() for ::highlight(foo)( should return an empty CSSStyleDeclaration
  assert_equals: expected "" but got "rgba(0, 0, 0, 0)"
[FAIL] getComputedStyle() for ::highlight(foo)(foo) should return an empty CSSStyleDeclaration
  assert_equals: expected "" but got "rgba(0, 0, 0, 0)"
[FAIL] getComputedStyle() for ::highlight(foo)() should return an empty CSSStyleDeclaration
  assert_equals: expected "" but got "rgba(0, 0, 0, 0)"
[FAIL] getComputedStyle() for :::highlight(foo) should return an empty CSSStyleDeclaration
  assert_equals: expected "" but got "rgba(0, 0, 0, 0)"
[FAIL] getComputedStyle() for ::highlight(foo). should return an empty CSSStyleDeclaration
  assert_equals: expected "" but got "rgba(0, 0, 0, 0)"
Harness: the test ran to completion.
